requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.29.0
django-ckeditor==6.5.1
dj-database-url==1.0.0
Django==4.0.6
django-heroku==0.3.1
django-social-share==2.3.0
environ==1.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.12
whitenoise==6.2.0
